added git status
