+++ client/src/components/Dog.js
@@ -6,7 +6,7 @@ import CardContent from "@material-ui/core/CardContent";
 import Button from "@material-ui/core/Button";
 import Typography from "@material-ui/core/Typography";
 import Divider from "@material-ui/core/Divider";
-import { useHistory } from "react-router-dom";
+import { useHistory, Link } from "react-router-dom";
 
 const useStyles = makeStyles({
   root: {
@@ -27,6 +27,7 @@ const useStyles = makeStyles({
 
 const Dog = ({ dogInfo, deleteDog }) => {
   const classes = useStyles();
+  const history = useHistory();
   return (
     <Card className={classes.root} variant="outlined">
       <CardContent>
@@ -43,7 +44,9 @@ const Dog = ({ dogInfo, deleteDog }) => {
         </Typography>
       </CardContent>
       <CardActions>
-        <Button variant="contained">EDIT</Button>
+        <Link to={`/editDog/${dogInfo.id}`}>
+          <Button variant="contained">EDIT</Button>
+        </Link>
         <Button variant="contained">GET INFO</Button>
         <Button
           variant="contained"
